Test Changes 1

+++ test/battle/ability/healer.c
@@ -3,7 +3,7 @@
 
 DOUBLE_BATTLE_TEST("Healer cures adjacent ally's status condition 30% of the time")
 {
-    u16 status;
+    u32 status;
     PARAMETRIZE { status = STATUS1_SLEEP; }
     PARAMETRIZE { status = STATUS1_POISON; }
     PARAMETRIZE { status = STATUS1_BURN; }
@@ -46,7 +46,7 @@ DOUBLE_BATTLE_TEST("Healer cures adjacent ally's status condition 30% of the tim
 
 DOUBLE_BATTLE_TEST("Healer cures status condition before burn or poison damage is dealt")
 {
-    u16 status;
+    u32 status;
     PARAMETRIZE { status = STATUS1_POISON; }
     PARAMETRIZE { status = STATUS1_BURN; }
     PARAMETRIZE { status = STATUS1_TOXIC_POISON; }
