+++ test/battle/ability/leaf_guard.c
@@ -4,7 +4,7 @@
 SINGLE_BATTLE_TEST("Leaf Guard prevents non-volatile status conditions in sun")
 {
     enum Move move;
-    u16 status;
+    u32 status;
     PARAMETRIZE { move = MOVE_WILL_O_WISP; status = STATUS1_BURN; }
     PARAMETRIZE { move = MOVE_HYPNOSIS; status = STATUS1_SLEEP; }
     PARAMETRIZE { move = MOVE_THUNDER_WAVE; status = STATUS1_PARALYSIS; }
@@ -36,7 +36,7 @@ SINGLE_BATTLE_TEST("Leaf Guard doesn't prevent non-volatile status conditions if
     enum Move move;
     u32 species;
     enum Ability ability;
-    u16 status;
+    u32 status;
     PARAMETRIZE { move = MOVE_WILL_O_WISP;  status = STATUS1_BURN;         species = SPECIES_GOLDUCK;  ability = ABILITY_CLOUD_NINE; }
     PARAMETRIZE { move = MOVE_HYPNOSIS;     status = STATUS1_SLEEP;        species = SPECIES_GOLDUCK;  ability = ABILITY_CLOUD_NINE; }
     PARAMETRIZE { move = MOVE_THUNDER_WAVE; status = STATUS1_PARALYSIS;    species = SPECIES_GOLDUCK;  ability = ABILITY_CLOUD_NINE; }
