+++ test/battle/item_effect/heal_and_cure_status.c
@@ -8,7 +8,7 @@ ASSUMPTIONS
 
 SINGLE_BATTLE_TEST("Full Restore restores a battler's HP and cures any primary status")
 {
-    u16 status;
+    u32 status;
     PARAMETRIZE { status = STATUS1_BURN; }
     PARAMETRIZE { status = STATUS1_FREEZE; }
     PARAMETRIZE { status = STATUS1_PARALYSIS; }
@@ -56,7 +56,7 @@ SINGLE_BATTLE_TEST("Full Restore restores a battler's HP and cures any primary s
 
 SINGLE_BATTLE_TEST("Full Restore restores a party members HP and cures any primary status")
 {
-    u16 status;
+    u32 status;
     PARAMETRIZE { status = STATUS1_BURN; }
     PARAMETRIZE { status = STATUS1_FREEZE; }
     PARAMETRIZE { status = STATUS1_PARALYSIS; }
@@ -107,7 +107,7 @@ SINGLE_BATTLE_TEST("Full Restore restores a party members HP and cures any prima
 
 SINGLE_BATTLE_TEST("Full Restore heals a battler from any primary status")
 {
-    u16 status;
+    u32 status;
     PARAMETRIZE { status = STATUS1_BURN; }
     PARAMETRIZE { status = STATUS1_FREEZE; }
     PARAMETRIZE { status = STATUS1_PARALYSIS; }
@@ -151,7 +151,7 @@ SINGLE_BATTLE_TEST("Full Restore heals a battler from any primary status")
 
 SINGLE_BATTLE_TEST("Full Restore heals a party member from any primary status")
 {
-    u16 status;
+    u32 status;
     PARAMETRIZE { status = STATUS1_BURN; }
     PARAMETRIZE { status = STATUS1_FREEZE; }
     PARAMETRIZE { status = STATUS1_PARALYSIS; }
